Improve the problem.

+++ core/world/ChunkTask.cpp
@@ -56,7 +56,7 @@ const glimmer::TileVector2D& glimmer::ChunkTask::GetChunkVertexCoordinates() con
 
 uint64_t glimmer::ChunkTask::GetFingerprint() const
 {
-    const uint64_t type = std::to_underlying(chunkType_) & 0x3ULL;
+    const uint64_t type = static_cast<uint64_t>(std::to_underlying(chunkType_)) & 0x3ULL;
     const uint64_t x = static_cast<uint64_t>(chunkVertexCoordinates_.x) & 0x7FFFFFFFULL;
     const uint64_t y = static_cast<uint64_t>(chunkVertexCoordinates_.y) & 0x7FFFFFFFULL;
     return type << 62 | x << 31 | y;
